requirements.txt
alpaca-py>=0.21.0
anthropic>=0.39.0
python-dotenv>=1.0.0
python-telegram-bot>=21.0

